Implement Pyrrhic to replace Fathom Bench: 23306418
