+++ Halogen/Halogen/Halogen.vcxproj
@@ -46,8 +46,18 @@
     <ClCompile Include="..\src\MoveGenerator.cpp" />
     <ClCompile Include="..\src\Network.cpp" />
     <ClCompile Include="..\src\Position.cpp" />
+    <ClCompile Include="..\src\Pyrrhic\tbchess.cpp">
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release No Popcnt|Win32'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release AVX2|Win32'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release No Popcnt|x64'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|x64'">true</ExcludedFromBuild>
+      <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release AVX2|x64'">true</ExcludedFromBuild>
+    </ClCompile>
+    <ClCompile Include="..\src\Pyrrhic\tbprobe.cpp" />
     <ClCompile Include="..\src\Search.cpp" />
-    <ClCompile Include="..\src\tbprobe.cpp" />
     <ClCompile Include="..\src\SearchData.cpp" />
     <ClCompile Include="..\src\TimeManage.cpp" />
     <ClCompile Include="..\src\TranspositionTable.cpp" />
@@ -66,10 +76,10 @@
     <ClInclude Include="..\src\MoveGenerator.h" />
     <ClInclude Include="..\src\Network.h" />
     <ClInclude Include="..\src\Position.h" />
+    <ClInclude Include="..\src\Pyrrhic\stdendian.h" />
+    <ClInclude Include="..\src\Pyrrhic\tbconfig.h" />
+    <ClInclude Include="..\src\Pyrrhic\tbprobe.h" />
     <ClInclude Include="..\src\Search.h" />
-    <ClInclude Include="..\src\stdendian.h" />
-    <ClInclude Include="..\src\tbconfig.h" />
-    <ClInclude Include="..\src\tbprobe.h" />
     <ClInclude Include="..\src\SearchData.h" />
     <ClInclude Include="..\src\TimeManage.h" />
     <ClInclude Include="..\src\TranspositionTable.h" />
