+++ Halogen/src/MoveGeneration.h
@@ -15,3 +15,22 @@ Move GetSmallestAttackerMove(const Position& position, Square square, Players co
 
 bool MoveIsLegal(Position& position, const Move& move);
 
+//Below code adapted with permission from Terje, author of Weiss.
+//--------------------------------------------------------------------------
+
+// Returns the attack bitboard for a piece of piecetype on square sq
+template <PieceTypes pieceType> inline
+uint64_t AttackBB(Square sq, uint64_t occupied = EMPTY)
+{
+	switch (pieceType)
+	{
+	case KNIGHT:	return KnightAttacks[sq];
+	case KING:		return KingAttacks[sq];
+	case BISHOP:	return BishopTable[sq].attacks[AttackIndex(sq, occupied, BishopTable)];
+	case ROOK:		return RookTable[sq].attacks[AttackIndex(sq, occupied, RookTable)];
+	case QUEEN:		return AttackBB<ROOK>(sq, occupied) | AttackBB<BISHOP>(sq, occupied);
+	default:		throw std::invalid_argument("piecetype is argument is invalid");
+	}
+}
+
+//--------------------------------------------------------------------------
