+++ Halogen/src/Pyrrhic/tbprobe.h
@@ -1,99 +1,43 @@
 /*
- * tbprobe.h
- * (C) 2015 basil, all rights reserved,
+ * Copyright (c) 2013-2020 Ronald de Man
+ * Copyright (c) 2015 Basil, all rights reserved,
+ * Modifications Copyright (c) 2016-2019 by Jon Dart
+ * Modifications Copyright (c) 2020-2020 by Andrew Grant
  *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
  *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
  *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
  * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 #ifndef TBPROBE_H
 #define TBPROBE_H
 
-#include "tbconfig.h"
-
-#ifdef __cplusplus
-extern "C"
-{
-#endif
-
-#ifndef TB_NO_STDINT
 #include <stdint.h>
-#else
-typedef unsigned char uint8_t;
-typedef unsigned short uint16_t;
-typedef unsigned uint32_t;
-typedef long long unsigned uint64_t;
-typedef char int8_t;
-typedef short int16_t;
-typedef int int32_t;
-typedef long long int64_t;
-#endif
-
-#ifndef TB_NO_STDBOOL
 #include <stdbool.h>
-#else
-#ifndef __cplusplus
-typedef uint8_t bool;
-#define true    1
-#define false   0
-#endif
-#endif
 
-/*
- * Internal definitions.  Do not call these functions directly.
- */
-extern bool tb_init_impl(const char *_path);
-extern unsigned tb_probe_wdl_impl(
-    uint64_t _white,
-    uint64_t _black,
-    uint64_t _kings,
-    uint64_t _queens,
-    uint64_t _rooks,
-    uint64_t _bishops,
-    uint64_t _knights,
-    uint64_t _pawns,
-    unsigned _ep,
-    bool     _turn);
-extern unsigned tb_probe_root_impl(
-    uint64_t _white,
-    uint64_t _black,
-    uint64_t _kings,
-    uint64_t _queens,
-    uint64_t _rooks,
-    uint64_t _bishops,
-    uint64_t _knights,
-    uint64_t _pawns,
-    unsigned _rule50,
-    unsigned _ep,
-    bool     _turn,
-    unsigned *_results);
+#include "tbconfig.h"
 
 /****************************************************************************/
 /* MAIN API                                                                 */
 /****************************************************************************/
 
-#define TB_MAX_MOVES                (192+1)
+#define TB_MAX_MOVES                256
 #define TB_MAX_CAPTURES             64
 #define TB_MAX_PLY                  256
-#define TB_CASTLING_K               0x1     /* White king-side. */
-#define TB_CASTLING_Q               0x2     /* White queen-side. */
-#define TB_CASTLING_k               0x4     /* Black king-side. */
-#define TB_CASTLING_q               0x8     /* Black queen-side. */
 
 #define TB_LOSS                     0       /* LOSS */
 #define TB_BLESSED_LOSS             1       /* LOSS but 50-move draw */
@@ -159,7 +103,7 @@ extern unsigned tb_probe_root_impl(
 /*
  * The tablebase can be probed for any position where #pieces <= TB_LARGEST.
  */
-extern unsigned TB_LARGEST;
+extern int TB_LARGEST;
 
 /*
  * Initialize the tablebase.
@@ -186,10 +130,6 @@ void tb_free(void);
  * PARAMETERS:
  * - white, black, kings, queens, rooks, bishops, knights, pawns:
  *   The current position (bitboards).
- * - rule50:
- *   The 50-move half-move clock.
- * - castling:
- *   Castling rights.  Set to zero if no castling is possible.
  * - ep:
  *   The en passant square (if exists).  Set to zero if there is no en passant
  *   square.
@@ -204,27 +144,12 @@ void tb_free(void);
  * - Engines should use this function during search.
  * - This function is thread safe assuming TB_NO_THREADS is disabled.
  */
-static inline unsigned tb_probe_wdl(
-    uint64_t _white,
-    uint64_t _black,
-    uint64_t _kings,
-    uint64_t _queens,
-    uint64_t _rooks,
-    uint64_t _bishops,
-    uint64_t _knights,
-    uint64_t _pawns,
-    unsigned _rule50,
-    unsigned _castling,
-    unsigned _ep,
-    bool     _turn)
-{
-    if (_castling != 0)
-        return TB_RESULT_FAILED;
-    if (_rule50 != 0)
-        return TB_RESULT_FAILED;
-    return tb_probe_wdl_impl(_white, _black, _kings, _queens, _rooks,
-        _bishops, _knights, _pawns, _ep, _turn);
-}
+unsigned tb_probe_wdl(
+    uint64_t white,     uint64_t black,
+    uint64_t kings,     uint64_t queens,
+    uint64_t rooks,     uint64_t bishops,
+    uint64_t knights,   uint64_t pawns,
+    unsigned ep,        bool     turn);
 
 /*
  * Probe the Distance-To-Zero (DTZ) table.
@@ -271,39 +196,20 @@ static inline unsigned tb_probe_wdl(
  * - This function is NOT thread safe.  For engines this function should only
  *   be called once at the root per search.
  */
-static inline unsigned tb_probe_root(
-    uint64_t _white,
-    uint64_t _black,
-    uint64_t _kings,
-    uint64_t _queens,
-    uint64_t _rooks,
-    uint64_t _bishops,
-    uint64_t _knights,
-    uint64_t _pawns,
-    unsigned _rule50,
-    unsigned _castling,
-    unsigned _ep,
-    bool     _turn,
-    unsigned *_results)
-{
-    if (_castling != 0)
-        return TB_RESULT_FAILED;
-    return tb_probe_root_impl(_white, _black, _kings, _queens, _rooks,
-        _bishops, _knights, _pawns, _rule50, _ep, _turn, _results);
-}
+unsigned tb_probe_root(
+    uint64_t white,    uint64_t black,
+    uint64_t kings,    uint64_t queens,
+    uint64_t rooks,    uint64_t bishops,
+    uint64_t knights,  uint64_t pawns,
+    unsigned rule50,   unsigned ep,
+    bool     turn,     unsigned *results);
 
-typedef uint16_t TbMove;
 
-#define TB_MOVE_FROM(move)                                                 \
-    (((move) >> 6) & 0x3F)
-#define TB_MOVE_TO(move)                                                   \
-    ((move) & 0x3F)
-#define TB_MOVE_PROMOTES(move)                                             \
-    (((move) >> 12) & 0x7)
+typedef uint16_t PyrrhicMove;
 
 struct TbRootMove {
-  TbMove move;
-  TbMove pv[TB_MAX_PLY];
+  PyrrhicMove move;
+  PyrrhicMove pv[TB_MAX_PLY];
   unsigned pvSize;
   int32_t tbScore, tbRank;
 };
@@ -325,21 +231,13 @@ struct TbRootMoves {
  *
  */
 int tb_probe_root_dtz(
-    uint64_t _white,
-    uint64_t _black,
-    uint64_t _kings,
-    uint64_t _queens,
-    uint64_t _rooks,
-    uint64_t _bishops,
-    uint64_t _knights,
-    uint64_t _pawns,
-    unsigned _rule50,
-    unsigned _castling,
-    unsigned _ep,
-    bool     _turn,
-    bool hasRepeated,
-    bool useRule50,
-    struct TbRootMoves *_results);
+    uint64_t white,    uint64_t black,
+    uint64_t kings,    uint64_t queens,
+    uint64_t rooks,    uint64_t bishops,
+    uint64_t knights,  uint64_t pawns,
+    unsigned rule50,   unsigned ep,
+    bool     turn,     bool hasRepeated,
+    bool useRule50,    struct TbRootMoves *results);
 
 /*
 // Use the WDL tables to rank and score all root moves.
@@ -353,47 +251,13 @@ int tb_probe_root_dtz(
  *   non-zero if ok, 0 means not all probes were successful
  *
  */
-int tb_probe_root_wdl(uint64_t _white,
-    uint64_t _black,
-    uint64_t _kings,
-    uint64_t _queens,
-    uint64_t _rooks,
-    uint64_t _bishops,
-    uint64_t _knights,
-    uint64_t _pawns,
-    unsigned _rule50,
-    unsigned _castling,
-    unsigned _ep,
-    bool     _turn,
-    bool useRule50,
+int tb_probe_root_wdl(
+    uint64_t white,    uint64_t black,
+    uint64_t kings,    uint64_t queens,
+    uint64_t rooks,    uint64_t bishops,
+    uint64_t knights,  uint64_t pawns,
+    unsigned rule50,   unsigned ep,
+    bool     turn,     bool useRule50,
     struct TbRootMoves *_results);
-
-/****************************************************************************/
-/* HELPER API                                                               */
-/****************************************************************************/
-
-/*
- * The HELPER API provides some useful additional functions.  It is optional
- * and can be disabled by defining TB_NO_HELPER_API.  Engines should disable
- * the HELPER API.
- */
-
-#ifndef TB_NO_HELPER_API
-
-extern unsigned tb_pop_count(uint64_t _bb);
-extern unsigned tb_lsb(uint64_t _bb);
-extern uint64_t tb_pop_lsb(uint64_t _bb);
-extern uint64_t tb_king_attacks(unsigned _square);
-extern uint64_t tb_queen_attacks(unsigned _square, uint64_t _occ);
-extern uint64_t tb_rook_attacks(unsigned _square, uint64_t _occ);
-extern uint64_t tb_bishop_attacks(unsigned _square, uint64_t _occ);
-extern uint64_t tb_knight_attacks(unsigned _square);
-extern uint64_t tb_pawn_attacks(unsigned _square, bool _color);
-
-#endif
-
-#ifdef __cplusplus
-}
-#endif
 
 #endif
